HSEARCH-495 IdentifierProducer: ClassCastException
git-svn-id: https://svn.jboss.org/repos/hibernate/search/trunk@19178 1b8cb986-b30d-0410-93ca-fae66ebed9b2

+++ hibernate-search/src/main/java/org/hibernate/search/MassIndexer.java
@@ -115,7 +115,7 @@ public interface MassIndexer {
 	 * @param maximum
 	 * @return
 	 */
-	MassIndexer limitIndexedObjectsTo(int maximum);
+	MassIndexer limitIndexedObjectsTo(long maximum);
 
 	/**
 	 * Starts the indexing process in background (asynchronous).
